+++ tests/testthat/test-stockDataDownload.R
@@ -2,7 +2,7 @@ context("Checking online data downloading")
 
 test_that("the embedded SP500 symbols are the same", {
   sink(file = tempfile())
-  stockData <- stockDataDownload(c("AAPL"), from = "2018-01-01", to = "2018-01-30")
+  stockData <- stockDataDownload(c("AAPL"), from = "2018-01-01", to = "2018-01-30", local_file = FALSE)
   sink()
   stockData <- as.numeric(lapply(stockData, mean))
   stockDataDownload_check <- quantmod::getSymbols("AAPL", from = "2018-01-01", to = "2018-01-30", auto.assign = FALSE)